# openapi.yaml
openapi: 3.0.3
info:
  title: WikiLivee API
  version: 3.0.0
  description: >
    Wiki-редактор с совместным редактированием в реальном времени.
    Страницы, таблицы (локальное хранение в PostgreSQL), backlinks,
    collaboration (WebSocket), версионирование, якорные комментарии,
    авторизация через JWT.
  license:
    name: MIT
servers:
  - url: https://balooai.ru
    description: Production
  - url: http://localhost:8080
    description: Локальная разработка
security:
  - bearerAuth: []
paths:
...
  /api/pages:
    get:
      operationId: listPages
      tags: [Pages]
      summary: Список всех страниц
      responses:
        '200':
          description: Список страниц
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PageSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: createPage
      tags: [Pages]
      summary: Создать страницу (author = текущий пользователь)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePageRequest'
      responses:
        '201':
          description: Страница создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/pages/{id}:
    get:
      operationId: getPage
      tags: [Pages]
      summary: Получить страницу
      parameters:
        - $ref: '#/components/parameters/pageId'
      responses:
        '200':
          description: Страница
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      operationId: savePage
      tags: [Pages]
      summary: Сохранить страницу (autosave). Создаёт снапшот в page_versions
      parameters:
        - $ref: '#/components/parameters/pageId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePageRequest'
      responses:
        '200':
          description: Обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdatePageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deletePage
      tags: [Pages]
      summary: Удалить страницу
      parameters:
        - $ref: '#/components/parameters/pageId'
      responses:
        '204':
          description: Удалено
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/pages/{id}/backlinks:
    get:
      operationId: getPageBacklinks
      tags: [Pages]
      summary: Страницы, которые ссылаются на данную (через wikiLink-блоки)
      parameters:
        - $ref: '#/components/parameters/pageId'
      responses:
        '200':
          description: Список обратных ссылок
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PageSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/pages/{id}/children:
    get:
      operationId: getPageChildren
      tags: [Pages]
      summary: Дочерние страницы
      parameters:
        - $ref: '#/components/parameters/pageId'
      responses:
        '200':
          description: Список дочерних страниц
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PageSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ════════════════════════════════════════════════════════
  # VERSIONING
  # ════════════════════════════════════════════════════════
  /api/pages/{id}/versions:
    get:
      operationId: getPageVersions
      tags: [Versioning]
      summary: История версий страницы
      parameters:
        - $ref: '#/components/parameters/pageId'
      responses:
        '200':
          description: Список версий, отсортированных по убыванию
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PageVersionSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/pages/{id}/versions/{version}:
    get:
      operationId: getPageVersion
      tags: [Versioning]
      summary: Получить конкретную версию
      parameters:
        - $ref: '#/components/parameters/pageId'
        - $ref: '#/components/parameters/version'
      responses:
        '200':
          description: Версия страницы
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageVersion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/pages/{id}/comments:
    get:
      operationId: getComments
      tags: [Comments]
      summary: Комментарии к странице (с якорными позициями)
      parameters:
        - $ref: '#/components/parameters/pageId'
      responses:
        '200':
          description: Список комментариев, отсортированных по дате создания
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: addComment
      tags: [Comments]
      summary: Добавить комментарий. anchorFrom/anchorTo — позиции ProseMirror
      parameters:
        - $ref: '#/components/parameters/pageId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
            example:
              author: Alice
              text: Нужно уточнить этот момент
              anchorFrom: 42
              anchorTo: 58
              anchorText: уточнить этот момент
      responses:
        '201':
          description: Комментарий создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      operationId: deleteComment
      tags: [Comments]
      summary: Удалить комментарий
      parameters:
        - $ref: '#/components/parameters/pageId'
        - name: commentId
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: UUID комментария
      responses:
        '204':
          description: Удалено
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # ════════════════════════════════════════════════════════
  # SEARCH
  # ════════════════════════════════════════════════════════
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT токен выдаётся при /api/auth/login или /api/auth/register.
        Срок действия — 30 дней.
  parameters:
    pageId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: TEXT ID страницы (hex-строка 32 символа)
...
    version:
      name: version
      in: path
      required: true
      schema:
        type: integer
      description: Номер версии
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Требуется авторизация (Bearer токен отсутствует или невалиден)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: page not found

    # ── Auth ─────────────────────────────────────────────
...
    PageSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        icon:
          type: string
          example: 📄
        parent_id:
          type: string
          nullable: true
          description: ID родительской страницы
        version:
          type: integer
        updatedAt:
          type: string
          format: date-time
    Page:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        icon:
          type: string
          example: 📄
        parent_id:
          type: string
          nullable: true
        content:
          type: array
          items:
            $ref: '#/components/schemas/Block'
        version:
          type: integer
        updatedAt:
          type: string
          format: date-time
    CreatePageRequest:
      type: object
      properties:
        title:
          type: string
          default: ''
        icon:
          type: string
          default: 📄
        parentId:
          type: string
          nullable: true
          description: ID родительской страницы
    UpdatePageRequest:
      type: object
      required: [title, content, version]
      properties:
        title:
          type: string
        content:
          type: array
          items:
            $ref: '#/components/schemas/Block'
        version:
          type: integer
          description: Текущая версия клиента (для оптимистичного контроля)
    UpdatePageResponse:
      type: object
      properties:
        id:
          type: string
        version:
          type: integer
          description: Новая версия после сохранения
    Block:
      type: object
      required: [id, type]
      properties:
        id:
          type: string
        type:
          type: string
          enum:
            - paragraph
            - heading
            - list
            - table_embed
            - page_link
        content:
          type: string
          nullable: true
          description: Текстовое содержимое блока
        props:
          type: object
          additionalProperties: true
          description: >
            Дополнительные свойства блока:
            heading → { level: 1|2|3, tiptapContent: [...] };
            table_embed → { tableId: string, tableName: string };
            page_link → { targetId: string, targetTitle: string };
            list → { ordered: boolean, tiptapContent: [...] }.

    # ── Versioning ───────────────────────────────────────
    PageVersionSummary:
      type: object
      properties:
        version:
          type: integer
        savedAt:
          type: string
          format: date-time
    PageVersion:
      type: object
      properties:
        pageId:
          type: string
        version:
          type: integer
        title:
          type: string
        content:
          type: array
          items:
            $ref: '#/components/schemas/Block'
        savedAt:
          type: string
          format: date-time

    # ── Comments ─────────────────────────────────────────
    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pageId:
          type: string
        author:
          type: string
        text:
          type: string
        anchorFrom:
          type: integer
          description: Начальная позиция выделения в ProseMirror (0 = нет якоря)
        anchorTo:
          type: integer
          description: Конечная позиция выделения в ProseMirror
        anchorText:
          type: string
          description: Снапшот выделенного текста на момент создания комментария
        createdAt:
          type: string
          format: date-time
    CreateCommentRequest:
      type: object
      required: [author, text]
      properties:
        author:
          type: string
        text:
          type: string
        anchorFrom:
          type: integer
          default: 0
          description: Начальная позиция в ProseMirror (0 = без якоря)
        anchorTo:
          type: integer
          default: 0
        anchorText:
          type: string
          default: ''
          description: Выделенный текст (снапшот)

    # ── Search ───────────────────────────────────────────
...